Feat: Add a variable to control whether face-today is applied
This variable gives me control on whether `cfw:face-today` is applied
to the whole day view or not. This is important because I actually
prefer the colors that `calfw-blocks` uses for rendering the calendar
than the default used by `emacs-calfw`, and setting the new
`cfw:hightlight-today` to nil lets me use it so.

This change has also been contributed back to `emacs-calfw` at:
kiwanami/emacs-calfw#154